+++ lib/gateway_api/internal/auth0/user.go
